reorganize files and delete schedule part
