+++ src/main/java/seedu/address/commons/core/PshMessages.java
@@ -1,6 +1,6 @@
-package seedu.address.commons.petCore;
+package seedu.address.commons.core;
 
-public class Messages {
+public class PshMessages {
     public static final String MESSAGE_UNKNOWN_COMMAND = "Unknown command";
     public static final String MESSAGE_INVALID_COMMAND_FORMAT = "Invalid command format! \n%1$s";
     public static final String MESSAGE_INVALID_PET_DISPLAYED_INDEX = "The pet index provided is invalid";
