--- src/main/java/seedu/address/logic/scheduleCommands/AddSlotCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.scheduleCommands;
-
-public class AddSlotCommand {
-}
