--- src/main/java/seedu/address/logic/scheduleCommands/CheckSlotCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.scheduleCommands;
-
-public class CheckSlotCommand {
-}
