--- src/main/java/seedu/address/logic/scheduleCommands/DeleteSlotCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.scheduleCommands;
-
-public class DeleteSlotCommand {
-}
