--- src/main/java/seedu/address/logic/scheduleCommands/EditSlotCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.scheduleCommands;
-
-public class EditSlotCommand {
-}
