--- src/main/java/seedu/address/logic/scheduleCommands/ShowSlotsCommand.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.scheduleCommands;
-
-public class ShowSlotsCommand {
-}
