--- src/main/java/seedu/address/logic/scheduleParser/DeleteSlotParser.java
@@ -1,4 +1,0 @@
-package seedu.address.logic.scheduleParser;
-
-public class DeleteSlotParser {
-}
