--- src/main/java/seedu/address/model/ReadOnlyUserPrefsPet.java
@@ -1,16 +1,0 @@
-package seedu.address.model;
-
-import java.nio.file.Path;
-
-import seedu.address.commons.core.GuiSettings;
-
-/**
- * Unmodifiable view of user prefs.
- */
-public interface ReadOnlyUserPrefsPet {
-
-    GuiSettings getGuiSettings();
-
-    Path getPetStoreHelperFilePath();
-
-}
