--- src/main/java/seedu/address/model/ScheduleModel.java
@@ -1,5 +1,0 @@
-package seedu.address.model;
-
-public interface ScheduleModel extends GeneralModel {
-
-}
